Refactor project into files
